@@ -29,6 +29,6 @@
     "tailwindcss": "^3.4.0"
   },
   "engines": {
-    "node": "18.x"
+    "node": "24.x"
   }
 }
