@@ -1,9 +1,9 @@
-const Stripe = require('stripe');
+import Stripe from 'stripe';
 
 // Initialize Stripe with the secret key from environment variables
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
 
-module.exports = async (req, res) => {
+export default async function handler(req, res) {
   if (req.method !== 'POST') {
     return res.status(405).json({ error: 'Method not allowed' });
   }
@@ -33,4 +33,4 @@ module.exports = async (req, res) => {
     console.error('Stripe Error:', error);
     res.status(500).json({ error: error.message });
   }
-};
+}
